1.  Integrated SMF's DOS MESS changes 2.  Added image_absolute_index() and image_from_absolute_index() calls 3.  [APPLE2] GfxDecode now used again for text

+++ mess/machine/a7800.c
@@ -129,7 +129,7 @@ UINT32 a7800_partialcrc(const unsigned char *buf, size_t size)
 
 	crc =(UINT32) crc32(0L,&buf[128],size-128);
 
-	logerror("A7800 Partial CRC: %08lx %d [%s]\n", (long) crc, size, &buf[1]);
+	logerror("A7800 Partial CRC: %08lx %d [%s]\n", (long) crc, (int)size, &buf[1]);
 
 	return crc;
 }
